Add BotController
